requirements.txt
fastapi
uvicorn
pymongo
dnspython
requests
beautifulsoup4
python-dotenv
jinja2
aiofiles
